# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Task Assistant Worker API
  description: |
    Cloudflare Workers + Hono 后端 HTTP API（PRD v1.2 / 技术方案 v1.4 对齐）。

    **鉴权**：多数 `/api/*` 需 `Authorization: Bearer <token>`。`token` 为 **JWT**（`POST /api/auth/login` 签发，`sub` = `users.id`）或开发用 **明文用户 id**（如 `local-dev-user`，须存在于 D1）。

    **SSE 对话**：`POST /api/chat/stream` 返回 `text/event-stream`；帧格式为 `event: <名称>` + `data: <JSON>` + 空行。事件类型见下方该路径说明及 `GET /health/sse-chat`。

    **管理 Prompt**：`/api/prompts/*` 需配置 Worker 环境变量 `ADMIN_API_SECRET`，并使用 `Authorization: Bearer <ADMIN_API_SECRET>` 或 `X-Admin-Token: <ADMIN_API_SECRET>`。
  version: 1.2.0
servers:
  - url: http://127.0.0.1:8787
    description: 本地 wrangler dev
  - url: https://{workerSubdomain}.workers.dev
    description: Cloudflare Workers 默认子域（部署后替换）
    variables:
      workerSubdomain:
        default: your-worker
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT（login）或明文 users.id（仅开发）
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        ai_nickname: { type: string }
        preferences:
          type: object
          additionalProperties: true
        created_at: { type: integer, format: int64 }
    UserPutRequest:
      type: object
      properties:
        name: { type: string }
        email: { type: string, nullable: true }
        preferences:
          type: object
          nullable: true
          additionalProperties: true
...
    PromptTemplate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        template_text: { type: string }
        scenario: { type: string }
        created_at: { type: integer, format: int64 }
paths:
...
  /api/user:
    get:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: 当前用户资料
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    put:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: 更新资料（name 全库唯一；email 可为 null）
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserPutRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
        '409':
          description: NAME_CONFLICT 或邮箱冲突
  /api/user/ai-name:
    put:
      tags: [User]
      security: [{ bearerAuth: [] }]
      summary: 更新 AI 助手昵称
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nickname]
              properties:
                nickname: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
  /api/tasks:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: 任务列表
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: project_id
          in: query
          schema: { type: string, nullable: true }
      responses:
        '200':
          description: 任务数组（结构见实现 taskRowToApi）
    post:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: 创建任务
      responses:
        '201':
          description: Created
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: 任务详情
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '404':
          description: 不存在
    put:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: 更新任务（含 detail / detail_json）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
    delete:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: 删除任务
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 已删除
...
  /api/prompts:
    get:
      tags: [PromptsAdmin]
      security: [{ bearerAuth: [], adminToken: [] }]
      summary: 列出 Prompt 模板（管理员）
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PromptTemplate' }
        '401':
          description: 未授权
        '503':
          description: ADMIN_NOT_CONFIGURED
    post:
      tags: [PromptsAdmin]
      security: [{ bearerAuth: [] }]
      summary: 创建模板
      responses:
        '201':
          description: Created
  /api/prompts/{id}:
    get:
      tags: [PromptsAdmin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromptTemplate' }
    put:
      tags: [PromptsAdmin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
    delete:
      tags: [PromptsAdmin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK